Fixes AudioFileBox

+++ packages/studio/forge-boxes/src/schema/std/AudioFileBox.ts
@@ -6,8 +6,8 @@ export const AudioFileBox: BoxSchema<Pointers> = {
     class: {
         name: "AudioFileBox",
         fields: {
-            1: {type: "int32", name: "start-in-seconds"},
-            2: {type: "int32", name: "end-in-seconds"},
+            1: {type: "float32", name: "start-in-seconds"},
+            2: {type: "float32", name: "end-in-seconds"},
             3: {type: "string", name: "file-name"}
         }
     }, pointerRules: {accepts: [Pointers.AudioFile, Pointers.FileUploadState], mandatory: true}
